+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [#1835](https://github.com/lapce/lapce/pull/1835): Add mouse keybinds
 - [#1856](https://github.com/lapce/lapce/pull/1856): Highlight git/VCS modified files in explorer, palette, and buffer tabs
 - [#1574](https://github.com/lapce/lapce/pull/1574): Panel sections can be expanded/collapsed
+- [#1938](https://github.com/lapce/lapce/pull/1938): Use dropdown for theme selection in settings
 
 ### Bug Fixes
 - [#1911](https://github.com/lapce/lapce/pull/1911): Fix movement on selections with left/right arrow keys